+++ Gamefiles/game.py
@@ -4,7 +4,7 @@ import time
 
 from Gamefiles.util import Util
 from Visuals.render import Render
-from Visuals.roadCreation import Road
+from Visuals.road import Road
 from Visuals.spriteCreation import Sprites
 
 import globals as gl
